VER QUIEN BORRA PAGOS

+++ php/borrar_pago.php
@@ -6,6 +6,7 @@ include('../php/superAdmin.php');
 $IdPago = $conn->real_escape_string($_POST['valorIdPago']);
 $IdCliente = $conn->real_escape_string($_POST['valorIdCliente']);
 $Tipo = $conn->real_escape_string($_POST['valorTipo']);
+$Motivo = $conn->real_escape_string($_POST['valorMotivo']);
 
 $fecha_corte = mysqli_fetch_array(mysqli_query($conn, 'SELECT * FROM clientes WHERE id_cliente='.$IdCliente));
 
@@ -22,25 +23,38 @@ if($area['area']!="Administrador"){
       echo '<script >M.toast({html:"Deuda Borrada.", classes: "rounded"})</script>';   
     }
   }
-  if(mysqli_query($conn, "DELETE FROM pagos WHERE id_pago = '$IdPago'")){
-    echo '<script >M.toast({html:"Pago Borrado.", classes: "rounded"})</script>'; 
-    if ($Tipo == 'Mensualidad') {
-      $ultimoPago =  mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM pagos WHERE id_cliente = $IdCliente AND tipo = 'Mensualidad' ORDER BY id_pago DESC LIMIT 1"));
-      $porciones = explode(" ", $ultimoPago['descripcion']);
-      $Mes = $porciones[0];
-      $Año = $porciones[1];
-      #ARREGLO EL CUAL DEFINE EL MES SIGUIENTE PARA LA FECHA DE CORTE DEL CLIENTE
-      $array =  array('ENERO' => '02','FEBRERO' => '03', 'MARZO' => '04','ABRIL' => '05', 'MAYO' => '06', 'JUNIO' => '07', 'JULIO' => '08', 'AGOSTO' => '09', 'SEPTIEMBRE' => '10', 'OCTUBRE' => '11', 'NOVIEMBRE' => '12',  'DICIEMBRE' => '01');      
-      $N_Mes = $array[$Mes];
-      #COMO ES DICIEMBRE ADELANTAMOS UN AÑO PORQUE LA FECHA DE CORTE YA ES EL SIGUINETE AÑO PAGO TODO DICIEMBRE
-      if ($Mes == 'DICIEMBRE') {  $Año ++; }
-      #FECHA DE CORTE SEGUN EL MES Y AÑO SELECCIONADO
-      $FechaCorte = date($Año.'-'.$N_Mes.'-05'); 
-      mysqli_query($conn, "UPDATE clientes SET fecha_corte='$FechaCorte' WHERE id_cliente='$IdCliente'");
-    }
+  $INFO_Pgo =  mysqli_fetch_array(mysqli_query($conn,"SELECT * FROM pagos WHERE id_pago = $IdPago"));
+  $Descripcion = $INFO_Pgo['tipo'].' :'.$INFO_Pgo['descripcion'];
+  $Cantidad = $INFO_Pgo['cantidad'];
+  $Realizo = $INFO_Pgo['id_user'];
+  $Tipo_Cambio = $INFO_Pgo['tipo_cambio'];
+  $Fecha_registro =$INFO_Pgo['fecha'].' '.$INFO_Pgo['hora'];
+  $HOY = date('Y-m-d');
+
+  if(mysqli_query($conn, "INSERT INTO pagos_borrados(cantidad, descripcion, realizo, tipo_cambio, fecha_hora_registro, motivo, borro, fecha_borrado) VALUES ('$Cantidad', '$Descripcion', $Realizo, '$Tipo_Cambio', '$Fecha_registro', '$Motivo', $id, '$HOY')")){
+
+    if(mysqli_query($conn, "DELETE FROM pagos WHERE id_pago = '$IdPago'")){
+      echo '<script >M.toast({html:"Pago Borrado.", classes: "rounded"})</script>'; 
+      if ($Tipo == 'Mensualidad') {
+        $ultimoPago =  mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM pagos WHERE id_cliente = $IdCliente AND tipo = 'Mensualidad' ORDER BY id_pago DESC LIMIT 1"));
+        $porciones = explode(" ", $ultimoPago['descripcion']);
+        $Mes = $porciones[0];
+        $Año = $porciones[1];
+        #ARREGLO EL CUAL DEFINE EL MES SIGUIENTE PARA LA FECHA DE CORTE DEL CLIENTE
+        $array =  array('ENERO' => '02','FEBRERO' => '03', 'MARZO' => '04','ABRIL' => '05', 'MAYO' => '06', 'JUNIO' => '07', 'JULIO' => '08', 'AGOSTO' => '09', 'SEPTIEMBRE' => '10', 'OCTUBRE' => '11', 'NOVIEMBRE' => '12',  'DICIEMBRE' => '01');      
+        $N_Mes = $array[$Mes];
+        #COMO ES DICIEMBRE ADELANTAMOS UN AÑO PORQUE LA FECHA DE CORTE YA ES EL SIGUINETE AÑO PAGO TODO DICIEMBRE
+        if ($Mes == 'DICIEMBRE') {  $Año ++; }
+        #FECHA DE CORTE SEGUN EL MES Y AÑO SELECCIONADO
+        $FechaCorte = date($Año.'-'.$N_Mes.'-05'); 
+        mysqli_query($conn, "UPDATE clientes SET fecha_corte='$FechaCorte' WHERE id_cliente='$IdCliente'");
+      }
+    }else{
+      echo "<script >M.toast({html: 'Ha ocurrido un error.', classes: 'rounded'});</script>";
+    } 
   }else{
-    echo "<script >M.toast({html: 'Ha ocurrido un error.', classes: 'rounded'});</script>";
-  }
+      echo "<script >M.toast({html: 'Ha ocurrido un error al insertar.', classes: 'rounded'});</script>";
+  } 
 }
 ?>
 <div id="mostrar_pagos">
